requirements.txt
Jinja2==2.10
click==6.7
PyYAML==5.1